Addition of Comments
Elaborated and refined comments in login.php and index.php.

Added top code blocks to the modules  which align with our scope -  for authoring, module purpose, and notes for update by team members as they work through the project.

These pages have existing .php files, but are just templates/copies from another page, and will need to be heavily revamped to implement the necessary functionality. The thought here was to provide a consistent starting point for the team.