Address Comment Block Issue
Discovered that although the previous commit tested successfully locally, it did not work when hosted on a webserver.

Issue resulted from the comment blocks at the top, which were formatted as HTML comments. As it turns out, a PHP session cannot start after ANY HTML has been executed - even if it is just a comment - in a non-local setting. Resolved and retested.